feat: separate hook to get extra target based on Airflow connection
