+++ tests/hooks/dbt/test_dbt_run.py
@@ -7,6 +7,8 @@ from airflow.exceptions import AirflowException
 from dbt.contracts.results import RunStatus
 from dbt.exceptions import DbtProfileError
 
+from airflow_dbt_python.hooks.dbt import DbtHook
+
 
 def test_dbt_run_task(hook, profiles_file, dbt_project_file, model_files):
     """Test a dbt run task."""
@@ -190,15 +192,15 @@ def test_dbt_run_task_compiled(
 
 
 def test_dbt_run_with_airflow_connection(
-    hook, dbt_project_file, model_files, airflow_conns, profiles_file
+    dbt_project_file, model_files, airflow_conns, profiles_file
 ):
     """Pulling a target from an Airflow connection."""
     for conn_id in airflow_conns:
+        hook = DbtHook(dbt_conn_id=conn_id)
         result = hook.run_dbt_task(
             "run",
             project_dir=dbt_project_file.parent,
             profiles_dir=profiles_file.parent,
-            target=conn_id,
             select=[str(m.stem) for m in model_files],
         )
 
@@ -213,18 +215,18 @@ def test_dbt_run_with_airflow_connection(
 
 
 def test_dbt_run_with_airflow_connection_and_no_profiles(
-    hook, dbt_project_file, model_files, airflow_conns
+    dbt_project_file, model_files, airflow_conns
 ):
     """Using an Airflow connection in place of a profiles file.
 
     We omit the profiles_file hook as it should not be needed.
     """
     for conn_id in airflow_conns:
+        hook = DbtHook(dbt_conn_id=conn_id)
         result = hook.run_dbt_task(
             "run",
             project_dir=dbt_project_file.parent,
             profiles_dir=None,
-            target=conn_id,
             select=[str(m.stem) for m in model_files],
         )
