+++ tests/operators/test_dbt_seed.py
@@ -287,7 +287,7 @@ def test_dbt_seed_with_airflow_connection(dbt_project_file, seed_files, airflow_
             task_id="dbt_task",
             project_dir=dbt_project_file.parent,
             select=[str(m.stem) for m in seed_files],
-            target=conn_id,
+            dbt_conn_id=conn_id,
         )
 
         execution_results = op.execute({})
@@ -295,7 +295,7 @@ def test_dbt_seed_with_airflow_connection(dbt_project_file, seed_files, airflow_
 
         assert run_result["status"] == RunStatus.Success
         assert op.profiles_dir is None
-        assert op.target == conn_id
+        assert op.dbt_conn_id == conn_id
 
 
 def test_dbt_seed_with_airflow_connection_and_profile(
@@ -306,14 +306,16 @@ def test_dbt_seed_with_airflow_connection_and_profile(
     An Airflow connection target should still be usable even in the presence of
     profiles file, and vice-versa.
     """
-    all_targets = airflow_conns + ("test",)
+    all_airflow_conns = airflow_conns + (None,)
+    target = "test"
 
-    for target in all_targets:
+    for conn_id in all_airflow_conns:
         op = DbtSeedOperator(
             task_id="dbt_task",
             project_dir=dbt_project_file.parent,
             profiles_dir=profiles_file.parent,
             select=[str(m.stem) for m in seed_files],
+            dbt_conn_id=conn_id,
             target=target,
         )
 
@@ -322,4 +324,4 @@ def test_dbt_seed_with_airflow_connection_and_profile(
 
         assert run_result["status"] == RunStatus.Success
         assert op.profiles_dir == profiles_file.parent
-        assert op.target == target
+        assert execution_results["args"]["target"] == conn_id or target
